requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
auth0-python==4.10.0
babel==2.17.0
backoff==2.2.1
backrefs==5.8
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
crewai==0.130.0
cryptography==45.0.4
cycler==0.12.1
Cython==3.0.11
dacite==1.9.2
databases==0.6.2
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
duckdb==1.3.1
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
filelock==3.18.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.58.4
frequency-controller==0.0.6
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.5
htmlmin==0.1.12
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
identify==2.6.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.9.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
json_repair==0.47.1
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
litellm==1.72.0
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
multimethod==1.12
multipledispatch==1.0.0
narwhals==1.43.1
natsort==8.4.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==1.26.4
oauthlib==3.3.1
objectpath==0.6.1
onnxruntime==1.22.0
openai==1.90.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
paginate==0.5.7
pandas==2.3.0
pandas_flavor==0.7.0
pantab==5.2.2
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pbr==6.1.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
phik==0.12.4
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
polars==1.31.0
posthog==5.4.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyjanitor==0.31.0
PyJWT==2.10.1
pymdown-extensions==10.15
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyvis==0.3.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyyaml_env_tag==1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
selectolax==0.3.30
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.41
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.46.0
sweetviz==2.3.1
sympy==1.14.0
tableauhyperapi==0.0.22502
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torequests==6.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uniparser==3.0.3
urllib3==2.5.0
uv==0.7.13
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
visions==0.8.1
watchdog==6.0.0
watchdogs==2.0.1
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wordcloud==1.9.4
xarray==2025.6.1
yarl==1.20.1
ydata-profiling==4.16.1
zipp==3.23.0
